Refactored to use PSR-4 compatible namespace
