+++ src/ExactTarget/ET_AssocArrayUtils.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace ExactTarget;
+
+class ET_AssocArrayUtils
+{
+    public static function isAssoc($array)
+    {
+        return ($array !== array_values($array));
+    }
+}
