+++ src/ExactTarget/ET_BaseObject.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace ExactTarget;
+
+class ET_BaseObject {
+	public  $authStub, $props, $filter, $organizationId, $organizationKey;
+	protected $obj, $lastRequestID;
+}
