+++ src/ExactTarget/ET_BounceEvent.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace ExactTarget;
+
+class ET_BounceEvent extends ET_GetSupport {
+	public  $getSinceLastBatch;
+	function __construct()
+	{
+		$this->obj = "BounceEvent";
+		$this->getSinceLastBatch = true;
+	}
+}
