+++ src/ExactTarget/ET_ClickEvent.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace ExactTarget;
+
+class ET_ClickEvent extends ET_GetSupport {
+	public  $getSinceLastBatch;
+	function __construct()
+	{
+		$this->obj = "ClickEvent";
+		$this->getSinceLastBatch = true;
+	}
+}
