+++ src/ExactTarget/ET_Email.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace ExactTarget;
+
+class ET_Email extends ET_CUDSupport {
+	public  $folderId;
+	function __construct()
+	{
+		$this->obj = "Email";
+		$this->folderProperty = "CategoryID";
+		$this->folderMediaType = "email";
+	}
+}
