+++ src/ExactTarget/ET_Folder.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace ExactTarget;
+
+class ET_Folder extends ET_CUDSupport {
+	function __construct() {
+		$this->obj = "DataFolder";
+	}
+}
