+++ src/ExactTarget/ET_List.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace ExactTarget;
+
+class ET_List extends ET_CUDWithUpsertSupport {
+	public  $folderId;
+	function __construct() {
+		$this->obj = "List";
+		$this->folderProperty = "Category";
+		$this->folderMediaType = "list";
+	}
+}
