+++ src/ExactTarget/ET_List_Subscriber.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace ExactTarget;
+
+class ET_List_Subscriber extends ET_GetSupport {
+	function __construct() {
+		$this->obj = "ListSubscriber";
+	}
+}
