+++ src/ExactTarget/ET_Organization.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace ExactTarget;
+
+class ET_Organization extends ET_CUDSupport {
+	function __construct() {
+		$this->obj = "Account";
+	}
+}
