+++ src/ExactTarget/ET_SDKUtils.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace ExactTarget;
+
+class ET_SDKUtils
+{
+    public static function getSDKVersion()
+    {
+	    return "FuelSDK-PHP-v0.9";
+    }
+}
