+++ src/ExactTarget/ET_SentEvent.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace ExactTarget;
+
+class ET_SentEvent extends ET_GetSupport {
+	public  $getSinceLastBatch;
+	function __construct()
+	{
+		$this->obj = "SentEvent";
+		$this->getSinceLastBatch = true;
+	}
+}
