+++ src/ExactTarget/ET_Subscriber.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace ExactTarget;
+
+class ET_Subscriber extends ET_CUDWithUpsertSupport {
+	function __construct() {
+		$this->obj = "Subscriber";
+	}
+}
