+++ src/ExactTarget/ET_UnsubEvent.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace ExactTarget;
+
+class ET_UnsubEvent extends ET_GetSupport {
+	public  $getSinceLastBatch;
+	function __construct()
+	{
+		$this->obj = "UnsubEvent";
+		$this->getSinceLastBatch = true;
+	}
+}
