+++ src/ExactTarget/ET_User.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace ExactTarget;
+
+class ET_User extends ET_CUDSupport {
+	function __construct() {
+		$this->obj = "AccountUser";
+	}
+}
